+++ CHANGELOG.md
@@ -2,6 +2,7 @@
 
 ### Added
 
+* Support for digitally signing with ECDSA keys
 * Support for digitally signing with DSA keys
 
 ### Changed